data_pipeline/requirements.txt
kafka-python==2.0.2
requests==2.26.0
python-dateutil==2.8.2
psycopg2==2.9.1
python-dotenv==0.19.0
pandas==1.3.3
nltk==3.6.3
matplotlib==3.4.3
seaborn==0.11.2
jupyterlab==3.1.14